refactor of circuits
