+++ src/bgg/circuits/mod.rs
@@ -1,7 +1,5 @@
 pub mod circuit;
-pub mod eval;
 pub mod gate;
 
 pub use circuit::*;
-pub use eval::*;
 pub use gate::*;
